test: Rename test_files to test_collector
The module collects the tests a run is given. It is about to collect
them from inside a file too, not only find the files, so name it for
what it does.
