+++ test/utils/CMakeLists.txt
@@ -40,10 +40,10 @@ target_sources(
     statetest_runner.cpp
     t8n.hpp
     t8n.cpp
+    test_collector.hpp
+    test_collector.cpp
     test_driver.hpp
     test_driver.cpp
-    test_files.hpp
-    test_files.cpp
     test_report.hpp
     test_report.cpp
     test_state.hpp
